QRCode_application/src/main/java/rs/ac/singidunum/qrcode/QRCodeReader.java
package rs.ac.singidunum.qrcode;

import java.io.FileInputStream;
import java.io.FileNotFoundException;
import java.io.IOException;

import javax.imageio.ImageIO;

import com.google.zxing.BinaryBitmap;
import com.google.zxing.ChecksumException;
import com.google.zxing.FormatException;
import com.google.zxing.NotFoundException;
import com.google.zxing.client.j2se.BufferedImageLuminanceSource;
import com.google.zxing.common.HybridBinarizer;

public class QRCodeReader {
	public static String readQRCode(String path) throws FileNotFoundException, IOException, NotFoundException, ChecksumException, FormatException {
		
		BinaryBitmap binaryBitmap = new BinaryBitmap(new HybridBinarizer(new BufferedImageLuminanceSource(ImageIO.read(new FileInputStream("test.png")))));
		String text = "";
		com.google.zxing.qrcode.QRCodeReader reader = new com.google.zxing.qrcode.QRCodeReader();
		text = reader.decode(binaryBitmap).getText();
		return text;
	}

}
